Add aliasing knowledge to `BitSlice`
This relatively small commit adds the aliasing awareness built up in the
previous commits to `BitSlice` methods and trait implementations,
allowing slices to carry alias-awareness in their ordinary work and
behave correctly throughout a program.

This will need testing and an example program to explicitly demonstrate
what the aliasing marker means, and how users can work with it, but the
bulk of the work ought to be done.

This commit would not be possible without the Rust compiler’s firm rules
and thorough type-checking capabilities.