Enable integers to unalias as themselves.
This change alters the `BitStore` associated type graph so that the
integers are themselves for `NoAlias`, rather than a `Cell`. This
makes the type instantiation more sensible for users, who may wish
to declare `let _: &BitSlice<O, u8>` instead of
`let _: &BitSlice<O, Cell<u8>>`.

`Cell` has been firmly set to be inert in the aliasing system, and
forbidden from crossing threads.

There is a note in `src/store.rs` explaining the reasons behind this
change and what was necessary to effect it.