Rename `Local` to `LocalBits`, as that is a very common word
This commit also refactors the `__bits_from_slice!` internal macro's
API in order to avoid a `rustfmt` error. Lines that are longer than
the limit and cannot be shrunk cause their subsequent block to
eternally indent. Since the `$store` parameter can be inferred by
the input slice and the output type slot, it is not necessary to
provide to the macro, and can be replaced by the `_` placeholder.

This change will be reflected in the `README` when it is rewritten.