Encode pointer mutability into the trait system
This system enables a type-checked guarantee that all `*const`-to-`*mut` raises
are done on pointers that had previously been lowered `*mut`-to-`*const`. This
property is checked by Miri and the UB optimizer, but these are slow to emit
diagnostics; the trait system has a faster error path for incorrect code.