Support `portable-atomic`
The `portable-atomic` crate is a third-party provider of atomic
behaviors on embedded architectures that may not necessarily have CPU
instructions for performing these operations.

This commit un-seals the `Radium` trait and implements it on the types provided
by `portable-atomic`. It does not attempt to do any other work to fit that crate
into the `radium` type families, or permit it to be a complete substitute for
the standard library atomics.

Our support for third-party crates is on a best-effort basis. We would prefer to
migrate these implementations into `portable-atomic` itself, as it does not
always produce the full atomic symbol suite and cannot expose that information
through `cfg` flags like the language distribution can.

Unfortunately, by committing to this dependency in the v1 series, migrating the
implementations out of `radium` and into `portable-atomic` requires cutting a v2
series to break the dependency cycle.